+++ Source/interfacegen/interfacegen-Bridging-Header.h
@@ -1,0 +1,14 @@
+//
+//  interfacegen-Bridging-Header.h
+//  sourcekitten
+//
+//  Created by Norio Nomura on 2/24/16.
+//  Copyright © 2016 SourceKitten. All rights reserved.
+//
+
+#ifndef interfacegen_Bridging_Header_h
+#define interfacegen_Bridging_Header_h
+
+#import "sourcekitd.h"
+
+#endif /* interfacegen_Bridging_Header_h */