+++ Source/sourcekitten/VersionCommand.swift
@@ -9,7 +9,7 @@
 import Commandant
 import Result
 
-private let version = "0.11.0"
+private let version = "0.12.0"
 
 struct VersionCommand: CommandType {
     let verb = "version"